raw_restful/requirements.txt
Flask==3.0.3
Jinja2==3.1.4
psycopg2-binary==2.9.9
